*big fat bug* solved: .mtime file was always updated
the .mtime file used to be updated even if the synchronization
failed. now .mtime works exactly as .db.txt, a .new one is created
by mddiff, and is up to smd-server to rename both.

it is still the case that the .mtime is updated only server side, since
mtime makes sense only locally to the server, and means nothing for the
client.

Conflicts:

	smd-server // because the --dry-run option