safely strip offlineimap headers
This program will strip the provided header recursively from all
emails in a Maildir folder. It is designed to replace the `find |
xargs sed` pipeline that would corrupt emails that had something like
the header in their body.

We add a small bit of documentation, the script's usage has more.

Closes: #1