implement xdelta3 and fix lua env replacement
This implements xdelta3 instead of xdelta. The two versions are
incompatible. While some Linux distributions, such as Debian,
provide xdelta and xdelta3 where both work according to their
version, other distributions, such as Fedora, only provide the
functionality of xdelta3 and symlink xdelta to xdelta3. This means
that syncmaildir can not run on those distributions. By switching
to xdelta3, most modern distributions should be supported.
The syntax on how to generate a delta file as well as how to apply
one is different in xdelta3 as well. This has been accounted for in
the change.

Replacing the lua env shebang in the Makefile was only working for
some of the lua files included in the project. This has been unified
to work for all lua files by altering the Makefile as well as the
two lua files smd-server and smd-client. The shebang now does not
contain a space after it (#!/usr/bin/env). That is arguably the
most common implementation, so the change accounts for that in the
Makefile too.